chore: Factor checking if caller is an admin
